Shooting done
Finished implementing the shooting feature, i.e. it calculates shots, hits, wounds, saves, and then damage. It should be stated however that the way the invuln save is currently being handled is less than ideal, so feel free to improve it as it currently only works if the target has both a normal and an invuln save which isnt always the case